Solver working for 3D variant
Algorithm is still too slow. Need to think of a way to separate regions
on horizontal planes and on diagonals.
Board is now implemented as a dictionary of tuples to colors. There is
maybe a better way to do this. Optimally, the pieces would be listed.
If only colors are kept, there is maybe a beter way to define the data
types.